Update to GWT 2.5.0 and Dagger 0.9.1
Also updates JUnit and plugins, and fix a version to maven-compiler-plugin,
as Maven 3 otherwise barks.

Also changes classifier=sources to type=java-source for javax.inject: we
don't need the sources JAR in the classpath, we unpack it in our
super-sources. type=java-source is not inherited transitively, and not
added to the classpath.

Finally, move super-sources to src/main/super and use a targetPath in the
POM.